ExceptionHelper WeakIdentityMap changed to ConditionalWeakTable. Performance characteristics unknown. CompilerclassLoader wasn't auto detecting and reloading base assembly. This preserves previous characteristics. But it's non-idea. This logic should come out of CompilerClassLoader. Split StaticCompiler out of CompilerClassLoader.
